requirements.txt
matplotlib
python-docx
pandas
